Move preview popover logic into own controller
